chg: move getReceiverAmount to utils
