+++ packages/router/src/lib/helpers/prepare.ts
@@ -2,15 +2,10 @@ import {
   AuctionBid,
   recoverAuctionBid as _recoverAuctionBid,
   decodeAuctionBid as _decodeAuctionBid,
-  calculateExchangeAmount,
-  getRateFromPercentage,
-  calculateExchangeWad,
+  getReceiverAmount as _getReceiverAmount,
 } from "@connext/nxtp-utils";
-import { BigNumber } from "ethers";
-
 import { AmountInvalid } from "../errors/prepare";
 
-const ROUTER_FEE = "0.05"; // 0.05%
 const EXPIRY_DECREMENT = 3600 * 24;
 const ONE_DAY_IN_SECONDS = 3600 * 24;
 const ONE_WEEK_IN_SECONDS = 3600 * 24 * 7;
@@ -30,19 +25,6 @@ export const validExpiryBuffer = (buffer: number) => buffer > ONE_DAY_IN_SECONDS
 export const validBidExpiry = (bidExpiry: number, currentTime: number) => bidExpiry > currentTime;
 
 /**
- * Returns the swapRate
- *
- * @param TODO
- * @returns The swapRate, determined by the AMM
- *
- * @remarks
- * TODO: getSwapRate using AMM
- */
-export const getSwapRate = async (): Promise<string> => {
-  return "1";
-};
-
-/**
  * Returns the amount * swapRate to deduct fees when going from sending -> recieving chain to incentivize routing.
  *
  * @param amount The amount of the transaction on the sending chain
@@ -59,14 +41,7 @@ export const getReceiverAmount = async (
   if (amount.includes(".")) {
     throw new AmountInvalid(amount);
   }
-  // 1. swap rate from AMM
-  const swapRate = await getSwapRate();
-  const amountAfterSwapRate = calculateExchangeWad(BigNumber.from(amount), inputDecimals, swapRate, outputDecimals);
-
-  // 2. flat fee by Router
-  const routerFeeRate = getRateFromPercentage(ROUTER_FEE);
-  const receivingAmount = calculateExchangeAmount(amountAfterSwapRate.toString(), routerFeeRate);
-  return receivingAmount.split(".")[0];
+  return await getReceiverAmount(amount, inputDecimals, outputDecimals);
 };
 
 /**
