+++ packages/utils/src/index.ts
@@ -22,3 +22,4 @@ export * from "./time";
 export * from "./gasEstimates";
 export * from "./fallbackSubgraph";
 export * from "./gelato";
+export * from "./router";
